Embed paginator within OfferService.
A paginator is a class that exists independently of its implementation, so it is okay to have it as an interface. Change name from PaginatorService to only Paginator, because Services p>

The implementation of a paginator for each class is directly linked to its class, it should not be publicly accesed by other classes. Made OfferPaginator private, inner, BUT not stayic with respect to OfferService. If other layers want to page offers, they can do so through the getPaginator method, which returns an unknown implementation of Paginator<Offer> interface.

@gbeade